+++ elemento.h
@@ -5,7 +5,7 @@ typedef struct Cliente{
     char *nombre;
     char *direccion;
     double telefono;
-    int costo;
+    float costo;
 }Cliente;
 
 typedef struct Repartidor{
@@ -17,16 +17,16 @@ typedef struct Repartidor{
 typedef struct Producto{
     struct Producto *ant, *sig;
     char *nombre;
-    int precio;
+    float precio;
     int existencias;
 }Producto;
 
 //Funciones Producto
-Producto *crearProducto(char *nombre, int precio, int existencias);
+Producto *crearProducto(char *nombre, float precio, int existencias);
 void imprimirProducto(Producto *e);
 
 //Funciones Cliente
-Cliente *nuevoCliente(char *nombre, char *direccion, double telefono, int costo);
+Cliente *nuevoCliente(char *nombre, char *direccion, double telefono, float costo);
 void imprimirCliente(Cliente * C);
 
 //FuncionesRepartidor
